Fix OpenTelemetry bugs: TraceContext handler, no-op payment spans, SDK version
- Convert TraceContext from Monolog Handler to Processor so trace_id/span_id
  are actually visible to subsequent handlers (was silently lost before)
- Remove no-op payment profiler spans in authorize/capture/refund/void that
  measured nothing (start+stop back-to-back with no work in between)
- Fix telemetry.sdk.version reporting actual SDK version instead of 'php'
- Cast ob_get_length() to int for proper OTel attribute typing
- Fix copyright years to 2026 on all new OpenTelemetry module files
- Restore original whitespace in Mage_Core_Model_Abstract

+++ app/code/core/Mage/Core/Model/App.php
@@ -371,7 +371,7 @@ class Mage_Core_Model_App
             if ($rootSpan) {
                 $rootSpan->setAttributes([
                     'http.status_code' => http_response_code(),
-                    'http.response_size' => ob_get_length() ?: 0,
+                    'http.response_size' => (int) (ob_get_length() ?: 0),
                 ]);
                 $rootSpan->setStatus('ok');
